requirements.txt
asgiref==3.4.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
dj-database-url==0.5.0
Django==3.2.9
django-ckeditor==6.1.0
django-jet-reboot==1.1.1
django-js-asset==1.2.2
djangorestframework==3.12.4
docx==0.2.4
feedparser==6.0.8
idna==3.3
lxml==4.6.4
Pillow==8.4.0
psycopg2==2.9.1
PyMySQL==1.0.2
python-docx==0.8.11
pytz==2021.3
reportlab==3.6.2
requests==2.26.0
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
whitenoise==5.3.0
..\venvs
